hosts/{atlas,luna,mars}: unify fan-control overlay node names
The Linux kernel thermal-zones devicetree documentation states specific
naming patterns to follow for naming the cooling-maps and thermal-trips
child nodes. We should follow this too.

Additionally, we will use the same fan curve for all devices for now,
but further testing is needed to determine if it is even worth having
different fan curves.