hosts/{luna,mercury,phobos,terra}: switch to root=gpt-auto
This allows us to avoid declaring (most of the) filesystem mounts in
NixOS and let systemd-gpt-auto-generator figure it all out for us during
stage-1: discovering and mounting the ESP, (LUKS encrypted) root and
swap partitions.

This commit also updates the installation instructions to reflect the
specific partitioning and installation steps required to enable gpt-auto
use.